Add create messsage and name header to desktop view
