+++ frontend/src/routes/messages/$conversationId.tsx
@@ -8,25 +8,14 @@ import { createFileRoute } from "@tanstack/react-router";
 import { useEffect, useState } from "react";
 import styled from "styled-components";
 import ProtectRoute from "../-preloaders/ProtectRoute";
+import CreateMessage from "./-components/CreateMessage";
+import MessageHeader from "./-components/MessageHeader";
 import { MessageList } from "./-components/MessageList";
 
 export const Route = createFileRoute("/messages/$conversationId")({
 	component: RouteComponent,
 	beforeLoad: ProtectRoute,
 });
-
-const DisplayName = styled.h2`
-  color: var(--sl-color-text);
-  text-align: center;
-  width: 50%;
-  margin: 0 0 10px 0;
-  font-size: 1.8rem;
-  font-weight: bold;
-
-  @media (max-width: 768px) {
-    font-size: 1.5rem;
-  }
-`;
 
 const Container = styled.div`
   display: flex;
@@ -48,7 +37,6 @@ const ChatContainer = styled.div`
   flex-direction: column;
   align-items: stretch;
   justify-content: flex-start;
-  overflow-y: auto;
   overflow-x: hidden;
   background: var(--sl-color-neutral-50);
   border-radius: 10px;
@@ -73,38 +61,6 @@ const ChatContainer = styled.div`
   }
 `;
 
-const SendMessageContainer = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: 1px solid var(--sl-color-primary-500);
-  border-radius: 20px;
-  margin-top: 20px;
-  background-color: var(--sl-input-background-color);
-
-  @media (max-width: 768px) {
-    width: 90%;
-    margin-right: 0;
-  }
-`;
-
-const MessageInput = styled(SlTextarea)`
-  width: 100%;
-  &::part(base) {
-    box-shadow: none;
-    border: 1px solid var(--sl-color-primary-500);
-    border-radius: 20px;
-    border: none;
-  }
-` as typeof SlTextarea;
-
-const ErrorMessage = styled.div`
-  color: var(--sl-color-danger-500);
-  font-size: 14px;
-  margin-top: 10px;
-`;
-
 function RouteComponent() {
 	const { conversationId } = Route.useParams();
 
@@ -118,9 +74,6 @@ function RouteComponent() {
 			path: { conversation_id: conversationId },
 		});
 
-	const [message, setMessage] = useState("");
-	const [isNoMessageBody, setIsNoMessageBody] = useState(false);
-
 	useEffect(() => {
 		const intervalId = setInterval(() => {
 			refetchConversation();
@@ -132,46 +85,9 @@ function RouteComponent() {
 		(conversationUser) => user?.id !== conversationUser.id,
 	);
 
-	const handleInputChange = (event: Event) => {
-		const target = event.target as HTMLInputElement;
-		setMessage(target.value);
-	};
-
-	const handleKeyDown = (event: Event) => {
-		const keyboardEvent = event as KeyboardEvent;
-		if (keyboardEvent.key === "Enter" && !keyboardEvent.shiftKey) {
-			keyboardEvent.preventDefault();
-			sendMessage();
-		}
-	};
-
-	const sendMessage = async () => {
-		if (!message) {
-			setIsNoMessageBody(true);
-			return;
-		}
-
-		try {
-			await createMessage({
-				body: { content: message },
-				path: { conversation_id: conversationId },
-			});
-		} catch (error) {
-			console.error(error);
-		}
-
-		setIsNoMessageBody(false);
-		setMessage("");
-		refetchConversation();
-	};
-
 	return (
 		<Container>
-			<DisplayName>
-				{otherUser
-					? `${otherUser.first_name} ${otherUser.last_name}`
-					: "An error occurred"}
-			</DisplayName>
+			<MessageHeader conversationId={conversationId} />
 			<ChatContainer>
 				{conversation ? (
 					<MessageList
@@ -182,34 +98,7 @@ function RouteComponent() {
 					<SlSpinner />
 				)}
 			</ChatContainer>
-			<SendMessageContainer>
-				<MessageInput
-					placeholder="Type a message..."
-					value={message}
-					spellCheck
-					rows={1}
-					resize="auto"
-					onSlInput={handleInputChange}
-					// @ts-ignore
-					onKeyDown={handleKeyDown}
-				/>
-				<SlIconButton
-					onClick={sendMessage}
-					name="send"
-					slot="suffix"
-					style={{
-						fontSize: "20px",
-						color: "var(--sl-color-text)",
-						marginRight: "5px",
-					}}
-				/>
-			</SendMessageContainer>
-			{isCreateMessageError && (
-				<ErrorMessage>Error sending message</ErrorMessage>
-			)}
-			{isNoMessageBody && (
-				<ErrorMessage>Message must contain text</ErrorMessage>
-			)}
+			<CreateMessage conversationId={conversationId} />
 		</Container>
 	);
 }
